+++ src/Graph.Lib/Indexes/DirectedAdjacencyMatrix.cs
@@ -1,25 +1,31 @@
 ﻿using Graph.Graphs;
+using System;
 
 namespace Graph.Indexes
 {
     public sealed class DirectedAdjacencyMatrix
         : AdjacencyMatrix
     {
-        public static IAdjacencyIndex<int> Empty { get; } = new DirectedAdjacencyMatrix();
+        public static IAdjacencyIndex<int> Empty()
+        {
+            return new DirectedAdjacencyMatrix();
+        }
 
         private DirectedAdjacencyMatrix()
             : base()
         {
         }
 
-        public DirectedAdjacencyMatrix(AdjacencyMatrix other) 
+        public DirectedAdjacencyMatrix(AdjacencyMatrix other)
             : base(other)
         {
         }
 
         public override bool Adjacent(int vertex1, int vertex2)
         {
-            return this.Matrix[vertex1, vertex2];
+            return vertex1 < this.Size
+                && vertex2 < this.Size
+                && this.Matrix[vertex1, vertex2];
         }
 
         public override object Clone()
@@ -31,7 +37,7 @@ namespace Graph.Indexes
         {
             if (vertex1 >= this.Size || vertex2 >= this.Size)
             {
-                this.Grow();
+                this.Grow(Math.Max(vertex1, vertex2));
             }
 
             if (!this.Matrix[vertex1, vertex2])
@@ -45,7 +51,7 @@ namespace Graph.Indexes
 
         public override bool Decouple(int vertex1, int vertex2)
         {
-            if (this.Matrix[vertex1, vertex2])
+            if (this.Adjacent(vertex1, vertex2))
             {
                 this.Matrix[vertex1, vertex2] = false;
                 return true;
