+++ src/Graph.Lib/Traversals/FastDepthFirstPreOrderTraversal.cs
@@ -19,15 +19,20 @@ namespace Graph.Traversals
 
         public override IEnumerable<int> Traverse(int vertex, int depth)
         {
+            var vertices = new int[this.AdjacencyIndex.Size];
             var visited = new bool[this.AdjacencyIndex.Size];
             var neighbors = new Stack<int>(new int[] { vertex });
+            var vi = 0;
 
             while (neighbors.Count > 0)
             {
                 var nextVertex = neighbors.Pop();
                 if (!visited[nextVertex])
                 {
-                    yield return nextVertex;
+                    //yield return nextVertex;
+                    vertices[vi] = nextVertex;
+                    ++vi;
+
                     visited[nextVertex] = true;
                     for (var i = this.AdjacencyIndex.Size - 1; i >= 0; --i)
                     {
@@ -40,6 +45,8 @@ namespace Graph.Traversals
                     }
                 }
             }
+
+            return vertices;
         }
     }
 }
